intel-sds-proto/sds/requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
anyjson>=0.3.3
argparse
Babel>=1.3
eventlet>=0.15.2
greenlet>=0.3.2
iso8601>=0.1.9
keystonemiddleware>=1.0.0
kombu>=2.5.0
lxml>=2.3
netaddr>=0.7.12
oslo.config>=1.4.0  # Apache-2.0
oslo.concurrency>=0.1.0  # Apache-2.0
oslo.db>=1.1.0  # Apache-2.0
oslo.messaging>=1.4.0
oslo.rootwrap>=1.3.0
oslo.serialization>=1.0.0               # Apache-2.0
oslo.utils>=1.0.0                       # Apache-2.0
osprofiler>=0.3.0                       # Apache-2.0
paramiko>=1.13.0
Paste
PasteDeploy>=1.5.0
pycrypto>=2.6
python-barbicanclient>=2.1.0,!=3.0.0
python-glanceclient>=0.14.0
python-novaclient>=2.18.0
python-swiftclient>=2.2.0
requests>=2.2.0,!=2.4.0
Routes>=1.12.3,!=2.0
taskflow>=0.4
rtslib-fb>=2.1.39
six>=1.7.0
SQLAlchemy>=0.9.7,<=0.9.99
sqlalchemy-migrate>=0.9.1,!=0.9.2
stevedore>=1.1.0  # Apache-2.0
suds>=0.4
WebOb>=1.2.3
wsgiref>=0.1.2
oslo.i18n>=1.0.0  # Apache-2.0
